server/pyproject.toml
[tool.poetry]
name = "brator-server"
version = "0.1.0"
description = ""
authors = ["Steinar V. Kaldager <steinarvkaldager@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
psycopg2 = "^2.8.6"
pydantic = "^1.8.1"
djangorestframework = "^3.12.2"
uvicorn = "^0.13.4"
gunicorn = "^20.0.4"
uvloop = "^0.15.2"
httptools = "^0.1.1"
whitenoise = "^5.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
coverage = {extras = ["toml"], version = "^5.5"}
pytest-django = "^4.1.0"
pytest-cov = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
